+++ src/lib/physics/2d/types.hpp
@@ -7,7 +7,7 @@ namespace GooBalls {
 namespace d2 {
 namespace Physics {
 
-typedef double TimeStep;
+typedef Float TimeStep;
 
 } // Physics
 } // d2
